Refactor api_client into topology_api_client for multiple Core classes
